+++ checkpoint/orbax/checkpoint/__init__.py
@@ -56,4 +56,4 @@ if _in_ipython_session:
 PyTreeCheckpointer = functools.partial(Checkpointer, PyTreeCheckpointHandler())
 
 # A new PyPI release will be pushed everytime `__version__` is increased.
-__version__ = '0.2.1'
+__version__ = '0.2.2'