+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2022-05-08
+
 ### Added
 - Option to allow users to disable automatic temporary directory cleanup upon
 CheckpointManager initialization.